backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
typer>=0.12.0
rich>=13.0.0
open-clip-torch>=2.26.0
torch>=2.4.0
torchvision>=0.19.0
insightface>=0.7.3
onnxruntime-gpu>=1.19.0
imagehash>=4.3.0
Pillow>=10.0.0
pillow-heif>=0.18.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
exifread>=3.0.0
scikit-learn>=1.5.0
pyyaml>=6.0
tqdm>=4.66.0
